media-source/URL-createObjectURL.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>URL.createObjectURL(mediaSource)</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
test(function() {
    var mediaSource = new MediaSource();
    var url = window.URL.createObjectURL(mediaSource);
    assert_true(url != null);
    assert_true(url.match(/^blob:.+/) != null);
}, "URL.createObjectURL(mediaSource) should return a unique Blob URI.");
</script>
</body>
</html>
